requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
dnspython==2.6.1
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-reflection==1.60.0
idna==3.7
importlib_resources==6.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
motor==3.4.0
multidict==6.0.5
pkgutil_resolve_name==1.3.10
protobuf==4.25.1
pymongo==4.7.2
PyMySQL==1.1.0
redis==5.0.4
referencing==0.35.1
rpds-py==0.18.1
ujson==5.9.0
yarl==1.9.4
zipp==3.18.1
